+++ src/osmscout/Searching.cpp
@@ -288,7 +288,7 @@ osmscout::Searching::Intersection Searching::SimulateNextCrossing(osmscout::Rout
     int panelHeight = 122;
 
     int scale = 100000;
-    double minX, minY;
+    long double minX, minY;
 
     minX = qMin(lastNode.lon, crossingNode.lon);
     minX = qMin(minX, waysNodes[0].lon);
